# src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: UTP CodeFest BFF API (workshop contract)
  version: 1.2.0
  description: |
    The full API the pre-built Next.js frontend expects from the Spring Boot BFF.
    The browser calls `/api/*` on the frontend's origin; Next.js proxies it to this service.

    Every error body has the shape `{ "error": "message" }`.
    Endpoints are built across the workshop: Health (Phase 1), Projects (Phase 2), Auth (Phase 3).

    **Implemented so far:** `GET /api/health`, `GET /api/projects`, `POST /api/projects`,
    `POST /api/auth/register`, `POST /api/auth/login`.
    Still to come: the JWT-protected `DELETE /api/projects/{id}` and Google sign-in.
    Switch to "Implemented (generated from code)" in the definition dropdown to see only live endpoints.
  license:
    name: MIT
servers:
  - url: /
    description: This server (wherever this spec is served from)
  - url: http://localhost:8080
    description: Local Spring Boot
...
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      operationId: register
      summary: Create a local account
      description: The password is stored as a BCrypt hash. The response never includes it, and no token is issued; log in next.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409":
          description: Username or email already registered
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              examples:
                username: { value: { error: Username is already taken } }
                email: { value: { error: Email is already registered } }
  /api/auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Log in with email or username
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Credentials valid; here is our JWT
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401":
          description: Same message for unknown account and wrong password, so attackers can't probe for accounts
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { error: Invalid email/username or password }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: The `token` from `/api/auth/login` or `/api/auth/google`.
  responses:
    BadRequest:
      description: Validation failed or the body isn't valid JSON
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: "title: Title is required" }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    User:
      type: object
      description: Never includes the password hash.
      required: [id, name, username, email, provider]
      properties:
        id: { type: string, example: 6ac0f95588e87dba324c74b9 }
        name: { type: string, example: Ali Hacker }
        username: { type: string, description: Stored lowercase, example: ali_hacker }
        email: { type: string, format: email, description: Stored lowercase, example: ali@utp.edu.my }
        provider: { type: string, enum: [local, google] }
        createdAt: { type: string, format: date-time }
        googleId: { type: [string, "null"], description: Google account ID (`sub`), google provider only }
        avatar: { type: [string, "null"], description: Google profile picture URL }
    RegisterRequest:
      type: object
      required: [username, email, password]
      properties:
        name: { type: string, maxLength: 60, description: Display name; defaults to the username, example: Ali Hacker }
        username:
          type: string
          pattern: "^[A-Za-z0-9._]{3,30}$"
          description: 3-30 letters, digits, dots or underscores; unique and case-insensitive
          example: ali_hacker
        email: { type: string, format: email, maxLength: 254, description: Unique and case-insensitive, example: ali@utp.edu.my }
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 72
          description: Hashed with BCrypt (cost 10). Capped at 72 because BCrypt ignores anything longer.
          example: CodeFest2026!
    LoginRequest:
      type: object
      description: Send `email` or `username` (one is enough). An `email` value without "@" is treated as a username.
      required: [password]
      properties:
        email: { type: string, example: ali@utp.edu.my }
        username: { type: string, example: ali_hacker }
        password: { type: string, format: password, example: CodeFest2026! }
      anyOf:
        - required: [email]
        - required: [username]
...
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token: { type: string, description: Our HMAC-signed JWT, example: eyJhbGciOiJIUzUxMiJ9... }
        user: { $ref: "#/components/schemas/User" }